Remove example highlighting
Highlighting of particular lines in examples is being removed.
This is since updating referenced line numbers is not a current priority. It may be re-added in the future, either manually or automatically.
It has been removed from blocks from both included files and page-specific blocks. This is so that consistency is maintained.