Create new country budget alignment guidance
Guidance agreed in here: https://discuss.iatistandard.org/t/guidance-on-implementing-aid-and-budget-alignment/831

The guidance will be added to the current IATI reference page under Overview for the time being. With the development of the new IATI website, all the guidance and consideration pages will be reviewed again and moved to the new IATI website.